Refactor test_pom_xml.py
